--- a/src/README.md
+++ b/src/README.md
@@ -7,11 +7,6 @@
 
 ## Development speedup tips
 
-Delete pages/editor.tsx and pages/groups.tsx (or have them return an empty
-component). These pages have a lot of dependencies that slow down webpack.
-
-On my personal computer, I have hot reload times of ~1.5 seconds.
-
 To make development even faster, you can take advantage of tailwind incremental
 builds (which don't seem to work with the default postcss setup for some
 reason).