--- a/src/pages/editor.tsx
+++ b/src/pages/editor.tsx
@@ -10,16 +10,32 @@
 // }
 
 import { PageProps } from 'gatsby';
-import React, { useEffect, useState } from 'react';
-import EditorPage from '../components/Editor/EditorPage';
+import React, { lazy, Suspense } from 'react';
+import { isDevelopment, lazyLoadConfig } from '../utils/lazyLoad';
+
+// Lazy load the EditorPage component
+const EditorPage = lazy(() => import('../components/Editor/EditorPage'));
 
 export default function EditorPageContainer(
   props: PageProps
 ): JSX.Element | null {
-  const [hasMounted, setHasMounted] = useState(false);
-  useEffect(() => {
+  const [hasMounted, setHasMounted] = React.useState(false);
+  
+  React.useEffect(() => {
     setHasMounted(true);
   }, []);
+
   if (!hasMounted) return null;
+
+  // In development mode, use lazy loading if enabled
+  if (isDevelopment && lazyLoadConfig.editor.enabled) {
+    return (
+      <Suspense fallback={lazyLoadConfig.editor.fallback()}>
+        <EditorPage {...props} />
+      </Suspense>
+    );
+  }
+
+  // In production or if lazy loading is disabled, load normally
   return <EditorPage {...props} />;
 }