--- a/src/utils/lazyLoad.tsx
+++ b/src/utils/lazyLoad.tsx
@@ -25,4 +25,4 @@ export const lazyLoadConfig = {
     enabled: isDevelopment,
     fallback: () => <div>Groups is loading...</div>
   }
-}; 
+};