_Place an "x" in the corresponding checkbox if it is done or does not apply to this pull request._

- [x] I have tested my code.
- [x] I have added my solution according to the steps [here](https://usaco.guide/general/adding-solution#steps).
- [x] I have followed the code conventions mentioned [here](https://usaco.guide/general/adding-solution/#code-conventions).
  - I understand that if it is clear that I have not attempted to follow these conventions, my PR will be closed.
  - If changes are requested, I will re-request a review after addressing them.
- [x] I have linked this PR to any issues that it closes.

I implemented lazy loading for heavy components for faster development speed. In the original src/README.md, it says:

"Delete pages/editor.tsx and pages/groups.tsx (or have them return an empty
component). These pages have a lot of dependencies that slow down webpack.

On my personal computer, I have hot reload times of ~1.5 seconds."

It seemed strange to be deleting files and altering the codebase, so I wrapped the editor.tsx and groups.tsx in lazy components to only load them when necessary. The page now has a hot reload of ~973 ms. The editor and groups components have an Largest Contentful Paint (LCP) score of ~0.91s, an Interaction to Next Paint (INP) of 129ms, and a CLS (Cumulative Layout Shift) score of 0.